requirements.txt
discord.py[voice]==2.1.0
ffmpeg==1.4
yt-dlp<2023.0.0,>=2022.10.4
spotdl==4.0.6
